gc of wal works

+++ terark-tools/sst_scan/sst_scan.hpp
@@ -27,7 +27,7 @@ inline std::string ToHex(const char* data_, int size_) {
   if (size_) {
     result.resize(2 * size_);
     auto beg = &result[0];
-    for (size_t i = 0; i < size_; ++i) {
+    for (int i = 0; i < size_; ++i) {
       unsigned char c = data_[i];
       beg[i * 2 + 0] = hextab[c >> 4];
       beg[i * 2 + 1] = hextab[c & 0xf];
